feat(functions): adopt knative handlers for admin
